+++ tests/test_polynomial_newton.py
@@ -15,7 +15,7 @@ from numpy.testing import assert_, assert_almost_equal
 
 from minterpy.global_settings import INT_DTYPE
 from minterpy.utils import eval_newton_polynomials
-from minterpy import Grid
+from minterpy import Grid, MultiIndexSet
 
 from minterpy import (
     NewtonPolynomial,
@@ -23,6 +23,11 @@ from minterpy import (
     CanonicalToNewton,
     NewtonToLagrange,
 )
+
+
+@pytest.fixture(params=["numpy", "numba", "numba-par"])
+def diff_backend(request):
+    return request.param
 
 
 def test_eval(MultiIndices, NrPoints, num_polynomials):
@@ -70,39 +75,129 @@ def test_eval_batch(MultiIndices, num_polynomials, BatchSizes):
     assert_almost_equal(yy_newton, yy_canonical)
 
 
-def test_partial_diff(SpatialDimension, PolyDegree, LpDegree):
-    newton_poly = build_random_newton_polynom(SpatialDimension, PolyDegree, LpDegree)
+class TestDiff:
+    """All tests related to the differentiation of polys. in the Newton basis.
+    """
 
-    # Check partial derivative on each dimension by comparing it with Canonical partial derivatives
-    for dim in range(SpatialDimension):
+    def test_zero_derivative(
+        self,
+        SpatialDimension,
+        PolyDegree,
+        LpDegree,
+        num_polynomials,
+        diff_backend,
+    ):
+        """Test taking the 0th-order derivative of polynomials."""
+        # Create a random Newton polynomial
+        newton_poly = build_random_newton_polynom(
+            SpatialDimension,
+            PolyDegree,
+            LpDegree,
+            num_polynomials,
+        )
+
+        # A derivative of order zero along all dimensions should be equivalent
+        # to the same polynomial
+        orders = np.zeros(SpatialDimension, dtype=INT_DTYPE)
+        zero_order_diff_newt = newton_poly.diff(orders)
+
+        # Assertion
+        assert_polynomial_almost_equal(zero_order_diff_newt, newton_poly)
+
+    def test_vs_canonical(
+        self,
+        SpatialDimension,
+        PolyDegree,
+        LpDegree,
+        num_polynomials,
+        diff_backend,
+    ):
+        """Test comparing the gradient with that computed in canonical basis.
+        """
+        # Create a random Newton polynomial
+        newton_poly = build_random_newton_polynom(
+            SpatialDimension,
+            PolyDegree,
+            LpDegree,
+            num_polynomials,
+        )
+
+        # Transform to the canonical basis
         trafo_n2c = NewtonToCanonical(newton_poly)
         canon_poly = trafo_n2c()
-        can_diff_poly = canon_poly.partial_diff(dim)
+
+        # Differentiate in the canonical basis and transform back
+        diff_order = np.ones(SpatialDimension, dtype=INT_DTYPE)
+        can_diff_poly = canon_poly.diff(diff_order)
         trafo_c2n = CanonicalToNewton(can_diff_poly)
         newt_can_diff_poly = trafo_c2n()
 
-        newt_diff_poly = newton_poly.partial_diff(dim)
+        # Differentiate the original polynomial
+        newt_diff_poly = newton_poly.diff(diff_order, backend=diff_backend)
 
+        # Assertion
         assert_polynomial_almost_equal(newt_can_diff_poly, newt_diff_poly)
 
-def test_diff(SpatialDimension, PolyDegree, LpDegree, num_polynomials):
-    newton_poly = build_random_newton_polynom(SpatialDimension, PolyDegree, LpDegree, num_polynomials)
+    def test_partial_diff(
+        self,
+        SpatialDimension,
+        PolyDegree,
+        LpDegree,
+        diff_backend,
+    ):
+        """Test taking the partial derivative of polynomials."""
+        # Create a random Newton polynomial
+        newton_poly = build_random_newton_polynom(
+            SpatialDimension,
+            PolyDegree,
+            LpDegree,
+        )
 
-    # A derivative of order zero along all dimensions should be equivalent to the same polynomial
-    zero_order_diff_newt = newton_poly.diff(np.zeros(SpatialDimension, dtype=INT_DTYPE))
-    assert_polynomial_almost_equal(zero_order_diff_newt, newton_poly)
+        # Check partial derivative on each dimension by comparing it
+        # with the partial derivative in the canonical basis
+        for dim in range(SpatialDimension):
+            # Transform to the canonical basis
+            trafo_n2c = NewtonToCanonical(newton_poly)
+            canon_poly = trafo_n2c()
+            # ...differentiate
+            can_diff_poly = canon_poly.partial_diff(dim)
+            # ...and transform back
+            trafo_c2n = CanonicalToNewton(can_diff_poly)
+            newt_can_diff_poly = trafo_c2n()
 
-    # Comparing the gradient with that computed in Canonical basis
-    trafo_n2c = NewtonToCanonical(newton_poly)
-    canon_poly = trafo_n2c()
-    diff_order = np.ones(SpatialDimension, dtype=INT_DTYPE)
-    can_diff_poly = canon_poly.diff(diff_order)
-    trafo_c2n = CanonicalToNewton(can_diff_poly)
-    newt_can_diff_poly = trafo_c2n()
+            # Differentiate the original polynomial
+            newt_diff_poly = newton_poly.partial_diff(
+                dim,
+                backend=diff_backend,
+            )
 
-    newt_diff_poly = newton_poly.diff(diff_order)
+            # Assertion
+            assert_polynomial_almost_equal(newt_can_diff_poly, newt_diff_poly)
 
-    assert_polynomial_almost_equal(newt_can_diff_poly, newt_diff_poly)
+    def test_unsupported_backend(
+        self,
+        SpatialDimension,
+        PolyDegree,
+        LpDegree,
+        num_polynomials,
+        diff_backend,
+    ):
+        """Test unsupported backend to differentiate Newton polynomials."""
+        # Create a Newton polynomial
+        mi = MultiIndexSet.from_degree(SpatialDimension, PolyDegree, LpDegree)
+        nwt_coeffs = np.random.rand(len(mi))
+        nwt_poly = NewtonPolynomial(mi, nwt_coeffs)
+
+        # Attempt to differentiate with a non-supported back-end
+        unsupported_backend = "numdumb"
+        with pytest.raises(NotImplementedError):
+            nwt_poly.partial_diff(0, backend=unsupported_backend)
+
+        with pytest.raises(NotImplementedError):
+            nwt_poly.diff(
+                order=np.ones(SpatialDimension, dtype=np.int_),
+                backend=unsupported_backend,
+            )
 
 
 def test_integrate_over_bounds_invalid_shape(